Refactoring: Extract Class: Extracted StackTrace methods from Logging
